+++ foxes/models/model_book.py
@@ -1,6 +1,6 @@
 import foxes.models as fm
 import foxes.variables as FV
-from foxes.tools import Dict
+from foxes.utils import Dict
 
 
 class ModelBook:
@@ -15,40 +15,40 @@ class ModelBook:
 
     Attributes
     ----------
-    point_models : foxes.tools.Dict
+    point_models : foxes.utils.Dict
         The point models. Keys: model name str,
         values: foxes.core.PointDataModel
-    rotor_models : foxes.tools.Dict
+    rotor_models : foxes.utils.Dict
         The rotor models. Keys: model name str,
         values: foxes.core.RotorModel
-    turbine_types : foxes.tools.Dict
+    turbine_types : foxes.utils.Dict
         The turbine type models. Keys: model name str,
         values: foxes.core.TurbineType
-    turbine_models : foxes.tools.Dict
+    turbine_models : foxes.utils.Dict
         The turbine models. Keys: model name str,
         values: foxes.core.TurbineModel
-    turbine_orders : foxes.tools.Dict
+    turbine_orders : foxes.utils.Dict
         The turbine orders. Keys: model name str,
         values: foxes.core.TurbineOrder
-    farm_models : foxes.tools.Dict
+    farm_models : foxes.utils.Dict
         The farm models. Keys: model name str,
         values: foxes.core.FarmModel
-    farm_controllers : foxes.tools.Dict
+    farm_controllers : foxes.utils.Dict
         The farm controllers. Keys: model name str,
         values: foxes.core.FarmController
-    partial_wakes : foxes.tools.Dict
+    partial_wakes : foxes.utils.Dict
         The partial wakes. Keys: model name str,
         values: foxes.core.PartialWakeModel
-    wake_frames : foxes.tools.Dict
+    wake_frames : foxes.utils.Dict
         The wake frames. Keys: model name str,
         values: foxes.core.WakeFrame
-    wake_superpositions : foxes.tools.Dict
+    wake_superpositions : foxes.utils.Dict
         The wake superposition models. Keys: model name str,
         values: foxes.core.WakeSuperposition
-    wake_models : foxes.tools.Dict
+    wake_models : foxes.utils.Dict
         The wake models. Keys: model name str,
         values: foxes.core.WakeModel
-    sources : foxes.tools.Dict
+    sources : foxes.utils.Dict
         All sources dict
 
     """
@@ -151,28 +151,35 @@ class ModelBook:
         slist = ["linear", "linear_amb", "quadratic", "quadratic_amb", "max", "max_amb"]
         for s in slist:
 
-            self.wake_models[f"Jensen_{s}"] = fm.wake_models.top_hat.JensenWake(
+            self.wake_models[f"Jensen_{s}"] = fm.wake_models.wind.JensenWake(
                 superposition=s
             )
-            self.wake_models[f"Jensen_{s}_k007"] = fm.wake_models.top_hat.JensenWake(
+            self.wake_models[f"Jensen_{s}_k007"] = fm.wake_models.wind.JensenWake(
                 k=0.07, superposition=s
             )
 
             self.wake_models[
                 f"Bastankhah_{s}"
-            ] = fm.wake_models.gaussian.BastankhahWake(superposition=s)
+            ] = fm.wake_models.wind.BastankhahWake(superposition=s)
             self.wake_models[
                 f"Bastankhah_{s}_k002"
-            ] = fm.wake_models.gaussian.BastankhahWake(k=0.02, superposition=s)
+            ] = fm.wake_models.wind.BastankhahWake(k=0.02, superposition=s)
+            
+            self.wake_models[
+                f"TurbOPark_{s}"
+            ] = fm.wake_models.wind.TurbOParkWake(superposition=s)
+            self.wake_models[
+                f"TurbOPark_{s}_k004"
+            ] = fm.wake_models.wind.TurbOParkWake(k=0.04, superposition=s)
 
         slist = ["ti_linear", "ti_quadratic", "ti_max"]
         for s in slist:
             self.wake_models[
                 f"CrespoHernandez_{s[3:]}"
-            ] = fm.wake_models.top_hat.CrespoHernandezTIWake(superposition=s)
+            ] = fm.wake_models.ti.CrespoHernandezTIWake(superposition=s)
             self.wake_models[
                 f"CrespoHernandez_{s[3:]}_k002"
-            ] = fm.wake_models.top_hat.CrespoHernandezTIWake(k=0.02, superposition=s)
+            ] = fm.wake_models.ti.CrespoHernandezTIWake(k=0.02, superposition=s)
 
         self.sources = Dict(
             name="sources",